refactor: change file permissions of theme/
Motivation
----------
This will ensure that UNIX file permissions of the created volume
`/themes` will be the default user and the default group of the host
machine.

Thus, no more access restrictions when editing the files in that folder
on the host.

See: https://stackoverflow.com/a/56844765
User with GID 1000 is apparently the first non-system user on a UNIX
system: https://www.redhat.com/sysadmin/user-account-gid-uid

How to test
-----------
1. `docker-compose down -v`
2. `rm -rf themes`
3. `docker-compose up`
4. `ls -lla`
5. Folder `./themes` should have the same file permissions as all the
   other folders